feat: Page editor refactor
