+++ src/features/ResourceManager/PageExplorer/DocumentEditorPlaceholder.tsx
